+++ app/src/foss/java/com/google/firebase/FirebaseApp.kt
@@ -1,0 +1,9 @@
+package com.google.firebase
+
+import android.content.Context
+
+class FirebaseApp private constructor() {
+    companion object {
+        fun initializeApp(context: Context): FirebaseApp? = null
+    }
+}