+++ app/src/foss/java/com/google/firebase/appcheck/FirebaseAppCheck.kt
@@ -1,0 +1,9 @@
+package com.google.firebase.appcheck
+
+class FirebaseAppCheck private constructor() {
+    fun installAppCheckProviderFactory(factory: Any) = Unit
+
+    companion object {
+        fun getInstance(): FirebaseAppCheck = FirebaseAppCheck()
+    }
+}