+++ app/src/foss/java/com/google/firebase/appcheck/debug/DebugAppCheckProviderFactory.kt
@@ -1,0 +1,7 @@
+package com.google.firebase.appcheck.debug
+
+class DebugAppCheckProviderFactory private constructor() {
+    companion object {
+        fun getInstance(): DebugAppCheckProviderFactory = DebugAppCheckProviderFactory()
+    }
+}