+++ app/src/foss/java/com/google/firebase/appcheck/playintegrity/PlayIntegrityAppCheckProviderFactory.kt
@@ -1,0 +1,7 @@
+package com.google.firebase.appcheck.playintegrity
+
+class PlayIntegrityAppCheckProviderFactory private constructor() {
+    companion object {
+        fun getInstance(): PlayIntegrityAppCheckProviderFactory = PlayIntegrityAppCheckProviderFactory()
+    }
+}