+++ app/src/foss/java/kotlinx/coroutines/tasks/Tasks.kt
@@ -1,0 +1,8 @@
+package kotlinx.coroutines.tasks
+
+import com.google.android.gms.tasks.Task
+
+suspend fun <T> Task<T>.await(): T {
+    exception?.let { throw it }
+    return result
+}